Move bastion probing method to FablibManager
